// src/app/components/model/ticket-reason-category.ts
export interface TicketReasonCategory {
  categoryName: any;
  ticketReasonCategoryTATMappingList: [];
  service: any;
  status: string;
  id: any;
  slaTimeP1: string;
  slaTimeP2: string;
  slaTimeP3: string;
  slaUnitP1: string;
  slaUnitP2: string;
  slaUnitP3: string;
  department: string;
  mvnoId: any;
  isDefaultProblemDomain: boolean;
  types: [];

}